Organize managers by category
