+++ Managers/Networking/WebSocketServerManager.cs
@@ -14,8 +14,9 @@ using System.Threading.Tasks;
 using Axys.Utilities;
 using Rhino.Commands;
 using Axys.Commands;
+using Axys.Managers.ObjectHandling;
 
-namespace Axys
+namespace Axys.Managers.Networking
 {
     public static class WebSocketServerManager
     {
