+++ Utilities/CommandUtilities.cs
@@ -5,6 +5,8 @@ using Rhino.Commands;
 using Rhino.DocObjects;
 using Rhino.Geometry;
 using Rhino.Input;
+using Axys.Managers.Networking;
+using Axys.Managers.ObjectHandling;
 
 namespace Axys.Utilities
 {
